Removes confusing and unnecessary explanation of HTTP basic auth
Instead of trying to teach users to do HTTP basic auth manually for API routes
that require it, this PR links to the article on the topic at the Mozilla
Developer Network and explains that, when using `curl` or a Conjur client
library, this is formatted for you automatically.

Additionally, this commit adds a `curl` example for the login route
demonstrating usage of basic auth.