More robust api.html generation
website.sh used to do docker-compose run to run a tool to
generate api.html. Unfortunately docker-compose will print
status output to stdout, so if the image needs to be built
the logs of that will end up in the generated file.

Instead of using `docker-compose run`, use docker-compose just
to build the image and run it using straight `docker run` for
full control over output.