Adds live-reload to Conjur website developer workflow
When you create a local docs site server using `docker-compose up
docs`, it now injects `livereload.js` which will automatically reload
the site the browser whenever it's regenerated.

As part of this effort, I resolved a weird issue with the placement of
Gemfile versus the assumed working directory inside the `conjur-docs`
container. It created a headache for me because you could run Jekyll
and generate the site from `/usr/src/docs` and it would pick up the
local plugins in `_plugins`, but wouldn't pick up plugins declared in
the Gemfile because it wouldn't see it.

So I changed the working directory to `/usr/src` and updated the
commands that generate the site to explicitly use `docs` as a source
and `docs/_plugins` as a plugin directory.