Optimize visibility queries
Unfortunately the optimizer isn't smart enough to push filter down
to use index in a recursive query, so we have to do it manually.
On the bright side, I used the opportunity to measure and optimize
more, and now most visibility queries take less than a millisecond
even for huge policies.