Update CONJUR_AUTHENTICATORS defaults per Kevin (more)
Kevin's quote from Slack:

I would rather see `CONJUR_AUTHENTICATORS` default to `authn`, but if
it’s defined explicitly, then the API key authenticator is not enabled
unless it’s specifically in the list. (edited) Because if I am in a
controlled environment like Kubernetes or AWS, I may want to literally
only permit the authenticator that is built for my environment.  This
reduces the threat surface, since API keys can’t be used as a backdoor.