Use more robust docker settings in security-scan
security-scan.sh ran as root in the containers, which
made it problematic re cleanup. Additionally it always
connected to network, even when it didn't need to.

This change abstracts out robust docker run configuration
into a shell lib, which is then used by security-scan.

Security-scan itself is also moved to ci directory, where
it rightfully belongs.