Tags published images with version number
This commit adds tagging to images based on Git tags. When the master
branch is built, if the Git release tag is the head of the master
branch, Docker images are created with the version number and pushed to
our internal repository, DockerHub, and Quay.io.

This change let's us directly tie a release to an image:
ex. `cyberark/conjur:1.1.1` matches the Git label `v1.1.1`.