Remove reference to jessie-updates
To test authn-k8s on GKE, we build an image based on the
google/cloud-sdk. The current version of that image has a reference to
the jessie-updates repo, which has been pulled from the distro.
Temporarily remove that reference, until we can determine whether we can
upgrade to a newer version of the cloud-sdk.