authn-k8s: Update cloud-sdk base image
The previously pinned image failed when attempting to fetch
the package index for the jessie-updates repo. I am not sure
why that operation was failing as the jessie-updates repo
still exists [1]. It may have been an incorrect deb entry
within the image.

This commit updates the cloud-sdk base image to the latest
version which doesn't show the same failure.

[1] http://ftp.us.debian.org/debian/dists/jessie-updates/