Upgrade docker base images
We've been using ubuntu 16.04 as the base images; that release is
nearing EOL, and the brighbox third-party repository that we've
been using to bring a more modern rubies into it has disappeared.

This patch changes the base images to ubuntu:18.04 and
phusion/baseimage:0.11 and removes the (now unnecessary)
discontinued brightbox ppa.