Not setting CONJUR_LOG_LEVEL to info by default
The log level is set by default by the rails environment. In is set in the environment setup to the value of CONJUR_LOG_LEVEL and if this variable
doesn't exist it is set to the environment's default  (debug for development, info for appliance & production). Having CONJUR_LOG_LEVEL set by default
to `info` may cause confusion for customers who change the RAILS_ENV variable to `development` and expect to see debug level logs.

To improve UX, now a user should explicitly add the CONJUR_LOG_LEVEL in order to set the log level to its value, rather than it appears "under the hood"
without the user noticing it.