Separate Bad Gateway test
Instead of having one scenario that tests both the dynamic change & the
Bad Gateway response we now have a scenario for each one, so the failure
is easily pin-pointed.

Also, we removed the commented-out part of the scenario and it will be added
as a new scenario when it is fixed.