Comment out shaky part of "provider-uri dynamic change" (#1155)
* Comment out shaky part of "provider-uri dynamic change"

In this test we test different values of provider-uri to verify:
  a. that a dynamic change of provider-uri is taken in run-time
  b. that different scenarios (502, 504) are running as designed (This is the integration test of these scenarios)

the test for the 504 response is shaky as "http://unreachable.com/" is not owned by us and does not always return a "504 Gateway Timeout" response. We should find a better solution for testing this.

Part a of this test is still good as we're checking the dynamic change

* Separate Bad Gateway test

Instead of having one scenario that tests both the dynamic change & the
Bad Gateway response we now have a scenario for each one, so the failure
is easily pin-pointed.

Also, we removed the commented-out part of the scenario and it will be added
as a new scenario when it is fixed.