Handle unset pod spec fields without error
Closes #1182

If `initContainers` or `containers` is not included in the deployment config for a pod, then the pod spec attributes for these are `nil` rather than an empty array.

This commit accounts for that possibility when searching for the authenticator containerby defaulting to an empty array if the value is `nil`. It also adds spec tests to verify this behavior.