Use advisory lock to ensure that only one server does the rotations.
This prevents race conditions or multiple update scenarios from happening.  Postgres
advisory locks take an integer and work as a db wide lock on that particular number.
This allows multiple Conjur servers to talk to the same database and perform rotations
without race conditions or conflicts.

This is needed because rotation is currently done on a batch process (retrieve
all variables that need rotated, then rotate each of them).  In the future it
would be nice to fix this so that only one variable to update is fetched at a time
(most-expired first?) allowing Conjur servers to cooperatively rotate variables
rather than one server doing all of the rotations each check.