Update the CSR's common-name in case of an old client
In order to authenticate hosts from every policy branch, we need the
authn-client to send the full host-id to the server. To maintain backwards
compatibility we need to support also client versions that do not send the
full host-id, and assume that the server adds the host-id prefix to it.

This commit updates the CSR in case of an old client so it will include
also the hard-coded prefix of the "apps" policy branch.

As part of this commit, we also:
- Add ability to update the common_name of SmartCsr
- Use full common-name to create the conjur host-id
  - This behavior replaces the current one that uses a hard-coded prefix
  with the 3-part suffix of the common-name