GKE coverage report is reliably generated
This commit removes the unnecessary `-it` flags to a kubectl exec
command for killing puma in the conjur container.

Conjur must be killed to ensure the simplecov report is written by
an at_exit hook. `-it` doesn't work in a Jenkins environment
as a tty isn't allocated for each executor.

Related: conjurinc/ops#470