Do not raise error in CSR namespace and spiffe id namespace difference
As part of the effort to define application identity in hosts'
annotations, the host id will no longer hold the id. Thus, we cannot
assume that the CSR (that has the host id) has the namespace inside it.
We still verify that the host's namespace matches the namespace in the
spiffe id but we do not extract the namespace from the CSR.