Allow defining application identity in host annotations
Prior to this commit, the application identity was defined only in the
host id. This commit enables users to define the application identity
in the host's annotations.

As part of this commit, the application identity is not extracted from
the CSR (that has the host id in it) but is extracted from either the
host's annotation or the host id.

Note: Application identity was called "Machine identity".