Set RAILS_ENV to test in ci
In our ci tests we have the Conjur server and Cucumber deployed
in containers. To be able to read the log file properly we need
to set RAILS_ENV to test in both containers as we have a shared
volume between the containers for reading the Conjur log in
the cucumber container.