Refactor AuthenticatorInput's password field
We would like to make this field more generic so it will
make sense in all of its uses. For example, in the authn-iam
authenticator, we use the AuthenticatorInput's password field
to extract the AWS header which are located in the request's
body. Calling that field `password` just because the body happens
to include the password in `authn` and `authn-ldap` is wrong.
This will also be the case in `authn-azure` where the body
will have the Azure access token and not a password.