Update NOTICES.txt to reflect proper formatting
These changes were requested by the legal team in order to enable
them to  review all the licenses and ensure we are using
'approved' gems. The format was made to match:
 https://github.com/cyberark/secretless-broker/blob/master/NOTICES.txt
as requested by the legal team.

Removed the gem 'random_password_generator', as it was not being used
and updated both the Gemfile and Gemfile.lock files to reflect this
update