Add test script for authn-azure
This script will deploy Conjur in an AWS machine, load policies into
Conjur (authenticator policy, azure-app & a test var). It will then
send authn-azure requests from the Azure VM to Conjur to get an access token
and use it to retrieve a secret.

This commit adds the part for Conjur (deploy & load policies).