Clean up code in decode_and_verify_id_token.rb
This commit makes the `call` method a bit clearer and moves
the logic of `ensure_certs_are_fresh` inside the `validate_id_token`
method (which is renamed to `decode_and_validate_id_token` as it tells
better its work.

The reason that `ensure_certs_are_fresh` is called inside the
`decode_and_validate_id_token` instead of in the `call` method is because
it is assumed that `ensure_certs_are_fresh` runs before `decode_and_validate_id_token`
for it to work (and is stated as so in the comments. Moving the method inside
`decode_and_validate_id_token` ensures that it will really run before the method's
logic, without the need to leave a note to future developers.