Improve code for OIDC authenticator
This commit does the following changes:
1. Use the `credentials` parameter of the `AuthenticatorInput` as we do
in other authenticators, instead of reading the request body by ourselves.
2. Remove the authenticate_request_body.rb class and move the logic into the
`authenticate` class. The logic also changes so that the output is a Hash, rather
than an object that has a hard-coded field `id_token`. This gives us more freedom
in adding more fields to the request body, and makes the solution more general.