Use `jwt` gem to verify and decode the token
We would like to have a generic class the verifies and decodes
JWT tokens. The way we decoded the token before this change was specific
for ID tokens and didn't handle access tokens. This commit makes it possible
to use this class in other authenticators that need to verify and decoded JWT tokens.

This commit also renames objects with `IDToken` to `Token` for generalization.