Collect coverage Reports pre merge
The conjur pipeline generates coverage reports in all of its branches,
most of which are merged together into a single report by simplecov
itself. This isn't possible for the gke report, so that is merged in
with jq. Recenly this has failed, this commit is for capturing more
information in order to determine why it's failing.

This is related to 15ee5b9ed4fc912bfd68c1565ef466b4bf80c24b which
moved artifact archival before merge, however that means only
the pre merge reports can be archived.

This commit ensures the source and merged reports are captured.
It also adds some explicit tests to the script so that if a report
is empty it will be noted in the console log.

Related: conjurinc/ops#516