Insert enabled_authenticators as input for ValidateStatus
The enabled_authenticators was inserted to ValidateStatus as a dependency
instead of as an input. This wasn't good as the enabled_authenticators can now
change in the DB so we need to load it each time the request is sent. Dependencies
are loaded when the server starts so if an authenticator is enabled in the DB so it
won't be reflected in the next time a request is sent. Moving to an input solves this as
the enabled_authenticators will be loaded for each request.