Update authenticator tests to run in isolation
The authenticator integration tests assumed that
the tests would run in-order and depending on state
from prior test scenarios.

These tests also shared environment and context with
the API test suite. So modifying the API tests to
run in isolation also requires modifying the authenticator
tests to run in isolation.

This commit makes the changes to run each scenario
in a fresh database context, and updates tests that
assumed database state from previous scenarios.