Make CredentialsController rspec tests pass under Rails 5
Code changes required by two breaking changes under Rails 5:

1. "ActionController::TestCase" is deprecated
=============================================

Rails 5 deprecates "ActionController::TestCase" and replaces it with
"ActionDispatch::IntegrationTest". See:

  https://blog.bigbinary.com/2016/04/19/changes-to-test-controllers-in-rails-5.html

for more information on that.

Confusing matters, this code interacts with these rails test classes
_via_ the rspec-rails gem, because we use rspec rather than Rails
recommended "Test::Unit". The underlying test class is specified with
the ":type" option, like so:

  describe AuthenticateController, :type => :controller do

which we needed to change to:

  describe AuthenticateController, :type => :request do

because the rspec-rails ":request" type maps to
"ActionDispatch::IntegrationTest", per the docs here:

  https://github.com/rspec/rspec-rails#what-tests-should-i-write

2. "RequestHelpers" changed signature of "post"/"get" helpers
=============================================================

This required:

- Manually constructing the URL endpoints we want to hit
- Changes to all invocations of "post" / "put" / "get"