Add trailing slash to Azure Provider URI in case it's missing
In case the user forgot to add a trailing slash to the provider-uri variable
(i.e `https://sts.windows.net/<tenant-id> instead of `https://sts.windows.net/<tenant-id>/`)
then they will get an error that is really hard to overcome. Instead, we can add the slash
for them in case they forgot it, without any harm.

This addition is not tested via UT as we mock the provider anyway. However, it does introduce a
new integration test that will be added once we have the infra for it (it is written and passes locally)