Add authn-azure tests to Jenkins pipeline
We have 2 Jenkins stages for the Azure Authenticator tests. one is responsible for allocating an Azure VM, from which we will get an Azure access token for Conjur authentication. It sets the Azure VM IP in the env and waits until the authn-azure tests are finished.

The other stage is responsible for running the tests. Once it finishes its run, it sets the env var `KEEP_AZURE_AUTHN_INSTANCE` to false which tells the first stage that it can finish its work and terminate the Azure VM.