Update SimpleCov timeout to include long test suites
SimpleCov has a timeout of 10 minutes for when the results
were writting to the .resultset.json file (so a suite that
finished at minute 1, one at minute 5 and one at minute 12
would only include the last two suites).  This sets that
window to 30 minutes, larger than should be necessary, but
short enough that it won't have any undue impact.