Remove low-impact possum references
Possum was an old codename used for the transition to open-source. It crept into
the source code in many places and has been a point of confusion ever since.
This commit removes all remaining low-impact references to possum. The rest of
the references are more difficult to remove and will require coordination across
other repositories.