Hosts can authenticate with Azure authenticator
This commit adds the application code for authenticating
hosts from Azure VMs according to their Azure access token.
The authenticator verifies the Azure access token against the
Azure Identity Provider, and if the token is valid it also validates
the Conjur host has the correct application identity stored in its
annotations, by comparing it to the data stored in the Azure access token.

This commit also adds a Status Check for the Azure Authenticator, so
Conjur operators can get immediate feedback on the authenticator
configuration. It also adds all the new log and error messages used
by the authenticator.