Azure Authenticator tests can run as part of ci/test
Add a new flag to the `ci/test` script for running the Azure
Authenticator tests. We will need some variables in the environment
in order to run the tests so we added the script check_dependencies.sh
that verifies all the variables are in the env. This file defines the
origin of each variable.