Run Azure Authenticator tests in Jenkins
We have 2 stages for the Azure Authenticator tests. One is
responsible for allocating an Azure VM, from which we will get an
Azure access token for Conjur authentication. It sets the Azure VM IP
in the env and waits until the authn-azure tests are finished. We use
this mechanism as we need a live Azure VM for getting the Azure token
and we don't want to have a long running-machine deployed in Azure.