Fixes support for deployment k8s authen for k8s 1.16 or newer
This change fixes support for using a deployment name as a Kubernetes
authenticator ID for Kubernetes server version 1.16 or newer.

This support had been broken for Kubernetes versions 1.16 or newer
because the server support for deployments (and some other Kubernetes
resources) using the deprecated api/v1beta1 endpoints was removed as
of Kubernetes 1.16, and the Kubernetes client that we use in the
Conjur Kubernetes authenticator was not updated.

The authenticator client had support for:
- Group: api (core), Version: v1
But was missing support for:
- Group: apis/apps, Version: v1
- Group: apis/extensions, Version: v1

The fix is to add explicit support for the missing apps/extensions
endpoints.

Fixes Issue #1440