Add log after fetching Identity Provider Keys
This log message was removed because I thought it is duplicated.
It actually isn't and tells us that the keys were retrieved from
the cache. The other message (`CONJ00009D`) will be written only
if we don't have the keys in the cache and we need to get them
from the provider.

Note that if we get the keys from the provider, both messages
will be written to the log.