Fix JSON cuke basic auth helper
This was the fix for an epic bug.

The original Slack discussion of its root cause can be found here:

https://conjurhq.slack.com/archives/CGTRB6G8G/p1582871801162500

Here is an edited version:

The correct behavior is that if the Authorization header is set, it
should  not get overridden.  But there was a bug in RestClient, fixed in
this diff:

https://github.com/rest-client/rest-client/commit/4ba9f7ee6374ec25abb870f769059b162ad7642a#diff-3724e1adc91ded3d8e979d73f91f1b1eL764

which allowed the Authorization header to get erronesouly overwritten,
because the test for an existing header was case sensitive.

So our code on master, at the time of this commit, is only working
because it’s relying on this bug.