Pin sequel below 5 to fix error
Prior to this commit, we were not explicitly including the sequel gem.

Instead, we got sequel indirectly via sequel-postgres-schemata, which is
a gem that Rafal wrote.  He recently updated the gem (pushed to gh march
13) with a commit which changed sequel requirement from ~> 4.3 to ">=
4.3", "< 6".

Regenerating Gemfile.lock during this work then causes sequel to bump
through a major version change, causing errors.

Sequel should be updated, but not as part of this work.  Instead, we fix
the lack of explicit dependency and this problem by pinning it to
version 4.49 in the Gemfile.

Slack discussion:

  https://conjurhq.slack.com/archives/CGTRB6G8G/p1584998714260000?thread_ts=1584986070.242400&cid=CGTRB6G8G