Remove cyberark/conjur-aws from landing page
At current, `cyberark/conjur-aws` is broken and needs more work to get
to a good state. We are reviewing whether AMIs are actually the recommended
path for deploying Conjur OSS to AWS, so it's possible we will not restore this
repo to its prior state. For now, we should remove it from the landing page.