Add default content type for Policy load API endpoints
If no content-type is specified in a POST request, Rack assumes the content
type is either `application/x-www-form-urlencoded` or
`multipart/form-data`.

The Conjur policy documents are YAML content, and some special
characters allowed in policy cause the Rack pody parsing to fail, leading to
a Rack error: `Invalid or incomplete POST parameters`.

This commit adds a Rack middleware that allow default content-types to be
configured on a per request path basis to better control the behavior when
no content-type is provided in the request headers.