Add a test case for resource list acting_as
During an effort to document the `acting_as` query parameter for the `resources`
route, it was raised that failing to specify the account in the fully qualified
ID of the role results in a 403. We may want to alter this behavior at some
point, but in the meantime it does not appear that we have a test that documents
this functionality. This commit adds a test, so that when/if we change this
behavior in the future, it will be clear that we have broken expected behavior
by breaking this test.