Log when using default authentication container name
If the annotation `authentication-container-name` is missing in
the host's annotations then we use a default value `authenticator`.

It can be helpful to the user to see this behaviour in the log. We already
log that an annotation was used to get this value
(e.g `Retrieved value of annotation {authn-k8s/authentication-container-name}`
but now we can also see that the default value is used.