Use ValidateSecurity class in ValidatePodRequest
We have a general class that handles the security validations for
authenticators (validate the webservice exists, it is enabled and the
user has permissions on it). This class was not used in ValidatePodRequest
and instead, it handled its own security. This was not ideal as some code was
duplicated and also it didn't perform some operations that are imeplemented in
ValidateSecurity (e.g raise a proper error if the account doesn't exist).

In addition, the spec of this file didn't inject the dependencies properly and
although they were defined in the spec, they were not really used. This commit fixes that.