Move log and error messages to centralized files (#1554)
* Move authn-iam logs and errors to the centralized files

We maintain our logs and errors in the centralized files
`logs.rb` & `errors.rb`. This commit moves those of the
IAM Authenticator.

We also removed the error log got the verification exception as it
will be printed to the log as part of the authentication process.

* Move schema log messages to centralized file

We maintain our logs and errors in the centralized files
`logs.rb` & `errors.rb`. This commit moves those of the
IAM Authenticator.

* Refactor RoleNotAuthorizedOnWebservice for generaliztion

We would like to use this error for other resources than webservices as well

* Move error to centralized files

We maintain our logs and errors in the centralized files
`logs.rb` & `errors.rb`. This commit moves those of the
IAM Authenticator.