Remove audit design doc
The OSS Conjur doesn't naturally make the syslog audit log available. This is
only meaningfully appliance in the DAP Appliance. The syslog format for audit
is documented for DAP elsewhere, so we remove it from the OSS docs.

The design doc also included development and test instructions for the audit
query engine. These are moved to the README in the audit engine directory.