Validate origin before credentials
It makes sense to validate the origin of the request before
validating the credentials. Before we validate that the credentials
are correct, we need to validate that the role _can_ authenticate,
same as we do with the security. We first verify that the request can
be authenticated and then validate it is authenticated.