Validate origin before credentials (#1564)
* Validate origin before credentials

It makes sense to validate the origin of the request before
validating the credentials. Before we validate that the credentials
are correct, we need to validate that the role _can_ authenticate,
same as we do with the security. We first verify that the request can
be authenticated and then validate it is authenticated.

* Use `def_delegators` for authenticator input

Instead of always calling @authenticator_input to get a value
of that object we can use `def_delegators`

* Use explicit input parameters instead of passing around `authenticator_input`

When we read a class definition, or somewhere that calls it, it is more readable to
see the actual inputs instead of some wrapper object (e.g `authenticator_input`)
that we need to dive into to see what it holds.