Raise proper error in case user not defined with `authn`
In case a user sends an authn request with the `authn` authenticator,
and the user is not defined in Conjur, the log message will show:
`Authentication Error: #<NoMethodError: undefined method `valid_origin?' for nil:NilClass>`.

This is not informative and doesn't tell the user the real story. This commit will change
the message to:
`Authentication Error: #<Errors::Authentication::Security::RoleNotFound: CONJ00007E ‘cucumber:user:non-existing-user’ not found>`