Use database query to filter role memberships by type
Previously each membership role was queried individually from the database to
check its "kind" and find the layers. This led to massive performance degredation
with a large number of memberships, as is the case with a host factory creating
many (100s) of hosts.

This commit implements that filter entirely in SQL to avoid the round trips and
improve performance at scale.