Refactor and simplify audit event infrastructure
The previous audit event classes were based on the unnecessary
Util::Struct class, and were full of confusing indirection: mutiple
levels of inheritance, as well as mixins (the CanFail module).

This made them difficult to understand, maintain, change, and test.

This commit contains the bulk of the work the removes the Util::Struct
dependency and all uses of inheritance and mixins.  Each class can now
be read and full understood on its own, with its full interface visible
in the class file itself.  Repeated code is handled through delegation,
in particular the Authn class, which the authentication events delegate
to.  Similarly, the CanFail module has been replaced by the
AttemptedAction class, an object we now delegate to rather than mixing
in.

We also simplify the interface itself for logging events, so that
callers only need to pass the logger a single parameter, the event
itself.

Finally, since we support two different loggers -- the Rails logger
which uses the standard ruby logger interface, as well as a syslog
logger -- we introduce adapter classes for each logger type, so that the
translation logic will be isolated.