Remove ValidateSecurity layer of abstraction
As part of the authentication process we validate that the authenticator
webservice is whitelisted and can be accessed by the requesting role.
These validations were encapsulated by ValidateSecurity but it made
the code less clear when reading the `call` methods of the command-classes
that were calling this validation.

Removing this layer, and calling the validations themselves will enhance the
readability of the authentication classes which are some of the main parts
 of the product.