Remove Err and Log definitions from top of classes
Each time we used an error class, or a log message class,
we would define its module at the top of the file. This was done
so when you call the class it would be concise without the whole path.

This approach led to issues where because we used memoization to init these
constants, they would sometimes mismatch what is written. For example,
if we had `Err ||= Errors::Authentication::AuthnAzure` in the top of the class
then sometimes the actual value of `Err` was `Errors::Authentication` because it was
already initialized with that value and the memoization meant that we don't initialize
it again.

A solution for that was to remove the memoization whenever this happened but then we had
another issue, where we got "already initialized constant" warnings. This happens because
now we initialize the same constant over and over again (e.g Err = Errors::Authentication::AuthnAzure).

Adding to the above the advantage of the explicit readability of calling error classes with their
full path led to the conclusion that we should remove the definition of `Err` and `Log` from the
top of classes, and call the full path each time.