Remove possible errors raised comments
We currently add a comment to each command class where we
specify the possible errors that are raised by this class.

This comment is hard-coded and it is the responsibility of the
developers to maintain it. The problem with this comment is that
it is not generated from anywhere and is not verified by code.

This leads us to a lot of cases where we move errors and forget to
update the comment so the comments are not accurate. Also, because
it is not verified then some classes have it and some classes do not.

It is better to not have any data at all than to have data that is incorrect,
thus we should remove these comments until they are forced by code of the
CommandClass interface.