Remove Err and Log definitions (#1645)
* Remove redundant definition check

We were defining the Errors and LogMessages modules after a guard-clause
but removing it had no impact

* Remove require of files under "app" directory

Rails auto-loader will load all the files under the "app"
directory so we don't need to explicitly require them

* Remove Err and Log definitions from top of classes

Each time we used an error class, or a log message class,
we would define its module at the top of the file. This was done
so when you call the class it would be concise without the whole path.

This approach led to issues where because we used memoization to init these
constants, they would sometimes mismatch what is written. For example,
if we had `Err ||= Errors::Authentication::AuthnAzure` in the top of the class
then sometimes the actual value of `Err` was `Errors::Authentication` because it was
already initialized with that value and the memoization meant that we don't initialize
it again.

A solution for that was to remove the memoization whenever this happened but then we had
another issue, where we got "already initialized constant" warnings. This happens because
now we initialize the same constant over and over again (e.g Err = Errors::Authentication::AuthnAzure).

Adding to the above the advantage of the explicit readability of calling error classes with their
full path led to the conclusion that we should remove the definition of `Err` and `Log` from the
top of classes, and call the full path each time.

* Remove possible errors raised comments

We currently add a comment to each command class where we
specify the possible errors that are raised by this class.

This comment is hard-coded and it is the responsibility of the
developers to maintain it. The problem with this comment is that
it is not generated from anywhere and is not verified by code.

This leads us to a lot of cases where we move errors and forget to
update the comment so the comments are not accurate. Also, because
it is not verified then some classes have it and some classes do not.

It is better to not have any data at all than to have data that is incorrect,
thus we should remove these comments until they are forced by code of the
CommandClass interface.

* Use rooted module of Util to avoid confusion

* Add memoization to remove warning

This commit removed "already initialized constant" warnings

* Break long lines