Save development.log as a Jenkins artifact
I was trying to figure out why a log entry wasn't matching a regex in a
cucumber test.  It was working locally.  I needed access to the
"development.log" on the CI.

This will come up again so we add it here as a permanent change.  It now
shows up in artifacts.