Validate required validations for default authenticator
Previously we didn't verify that the user exists in authentication requests
with the default authenticator. However, verifying this can lead to enhanced
log supportability as in case a user doesn't exist it is better to log
a `RoleNotFound` message rather than an `InvalidCredentials` message.

Note: this message will not be returned in the response and will only be written
to the log.