Pass role_id instead of role to authn events
All of these events only ever use the "role.id" method of the "role"
object that is passed to them, so passing the entire "role" object is
unnecessary and misleading.

We also introduce an Audit::Event::Authn::RoleId to encapsulate the
conditional logic around the meaning of "role id" in the context of
authentication events.