Use generic authentication code in authn-oidc
Prior to this commit, the OIDC authenticator implemented all the
general authentication process (webservice & user validations,
token generation, audit, etc.) whereas the rest of the authenticators
used a general authentication class that performed these actions for them.

This helps us in maintaining the code so that if a change is made somewhere
(e.g how we audit success) then we need to update it in one place instead of
in all the authenticator classes.

The reason why the OIDC authenticator implemented everything by itself is because
it gets the username from the provided ID token, and not from the URL like the rest
of the authenticators do.

So in order to use the general authentication class we first need to decode the ID token
and extract the username from it, and then use the general class.