Refactor application identity (#1580)
Refactor Application Identity code

This commit aligns the code of Application Identity validations between authenticators. It does not generalize the validations at this phase, and it only refactors the code of the validations in authn-k8s and authn-azure to better demonstrate the similarities between the two. After this effort is done we may generalize the validations but this step is a crucial one towards that. It also has enough value on its own as after this change it will be easier for future developers to add application identity validations for a new authenticator.

This commit also changes the term "Application Identity" to "Resource Restrictions" as it better tells the story. Although the term Application Identity is documented in our official docs, it is not clear to readers of the code and "ResourceRestrictions" are more self explanatory, as that is what application identity is all about. It defines the resources from which an authentication request is restricted to (like we have with IP restrictions).