Re-use authenticator_input for authn-oidc
It is best to have only one `authenticator_input` method so that in case
it needs to be changed (e.g add another input parameter) then we don't need
to do the change twice (or maybe miss it).

The only difference between the two is that the authenticator name is not in `params`
and that the `username` is nil. Both are handled in this PR.