Remove the ActionDispatch:RemoteIp middleware
This middleware provides additional logic for selecting the client IP address in Conjur,
and stores the client IP in `request.remote_ip`. This can sometimes be different than
the client IP determined by Rack in `request.ip`.

We use `request.ip` for audit and CIDR restrictions. However, Rails uses `request.remote_ip` in the request log.

In order to make the request log match the audit log and network restrictions, we remove this middleware. This causes `request.remote_ip` to fallback to the value in `request.ip`.