Only trust `127.0.0.1` by default to send the `X-Forwarded-For` header
Previously, if `TRUSTED_PROXIES` was not configured, Conjur would fallback to
the default Rack behavior for which IP addresses to trust to send the
`X-Forwarded-For` HTTP header. This used a regular expression to determine whether an IP address was routable or not. Any IP address considered non-routable according to this regex was trusted.

This commit modifies this behavior to only trust `127.0.0.1` by default, effectively locking down the `X-Forwarded-For` header, unless explicitly trusted with the
`TRUSTED_PROXIES` environment variable.