Make credentials get 'Any' instead of 'String'
In authn-gcp the credentials are a JWT token and we
replace it with the decoded token once we decode it.
This is done because the token will be needed later to
validate the Resource Restrictions and we don't want to
decode the token twice.

Thus, the type of 'credentials' should not be a String