Run authn-azure tests in an azure node
Prior to this commit we allocated an Azure node that just exported its
IP to the environment and then waited to the actual tests to finish.

This commit changes the behaviour so that the tests will run on the Azure
node itself and we have one less node running.

It also simplifies the Jenkinsfile