Run GCE Authenticator tests in Jenkins
We have 2 stages for GCE Authenticator tests. The first one runs inside
a GCE instance and retrieves all the tokens that will be used in the tests.
It then stashes the tokens, which are unstashed in the stage that runs the
GCE Authenticator tests using the tokens.
This way we can have a light-weight GCE instance that has no need for conjurops
or git identities and is not open for SSH