Remove cert file from correct pod
Before each cucumber test of authn-k8s, we delete the cert file
from `/etc/conjur/ssl` so that each test starts with a fresh folder.

For some reason we skipped the pod if its name started with `conjur-authn-k8s`
(which is the pod of the Conjur server) and performed the erase from pods that
do not start with this prefix. However, we need to erase the cert only from the
pod that performs the authentication which has the `inventory-deployment` prefix