Use explicit Ruby version in Gemfile
Fuzzy version matching (~>) for the Ruby version in the Gemfile does not play
nicely with RVM. We should be explicit about our Ruby version since it's such a
fundamental part of a Rails project. The Ruby version is also locked in place by
the Docker base image so it won't be updated with fuzzy matching.